Introduce TagUtils.getDistinctTags for reuse between resource types
Reorganize unit tests.
