+++ src/test/java/bio/terra/service/dataset/DatasetJsonConversionTest.java
@@ -1,9 +1,6 @@
 package bio.terra.service.dataset;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.contains;
-import static org.hamcrest.Matchers.containsInAnyOrder;
-import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -28,7 +25,6 @@ import bio.terra.model.TableModel;
 import bio.terra.service.resourcemanagement.MetadataDataAccessUtils;
 import bio.terra.service.resourcemanagement.google.GoogleProjectResource;
 import java.time.Instant;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.UUID;
@@ -292,60 +288,5 @@ public class DatasetJsonConversionTest {
     assertTrue("Primary key columns are marked as required", pkColumn.isRequired());
     assertFalse("Regular columns default to not required", defaultColumn.isRequired());
     assertTrue("Required columns are marked as required", requiredColumn.isRequired());
-  }
-
-  @Test
-  public void testDatasetRequestToDatasetTags() {
-    var requestNullTags = createBaseRequest().tags(null);
-    assertThat(
-        "Null tag list is converted to empty list",
-        DatasetJsonConversion.datasetRequestToDataset(requestNullTags).getTags(),
-        empty());
-
-    var requestEmptyTags = createBaseRequest().tags(List.of());
-    assertThat(
-        "Empty tag list is returned",
-        DatasetJsonConversion.datasetRequestToDataset(requestEmptyTags).getTags(),
-        empty());
-
-    List<String> nullTagElements = new ArrayList<>();
-    nullTagElements.add(null);
-    var requestNullTagElements = createBaseRequest().tags(nullTagElements);
-    assertThat(
-        "Null tags are filtered out",
-        DatasetJsonConversion.datasetRequestToDataset(requestNullTagElements).getTags(),
-        empty());
-
-    String tag = "a tag";
-    var requestDuplicateTags = createBaseRequest().tags(List.of(tag, tag, tag));
-    assertThat(
-        "Duplicate tags are filtered out",
-        DatasetJsonConversion.datasetRequestToDataset(requestDuplicateTags).getTags(),
-        contains(tag));
-
-    List<String> multipleCasedTags = List.of(tag.toLowerCase(), tag.toUpperCase());
-    var requestMultipleCasedTags = createBaseRequest().tags(multipleCasedTags);
-    assertThat(
-        "Tags are case sensitive",
-        DatasetJsonConversion.datasetRequestToDataset(requestMultipleCasedTags).getTags(),
-        containsInAnyOrder(multipleCasedTags.toArray()));
-  }
-
-  private DatasetRequestModel createBaseRequest() {
-    return new DatasetRequestModel()
-        .name(DATASET_NAME)
-        .defaultProfileId(DATASET_PROFILE_ID)
-        .description(DATASET_DESCRIPTION)
-        .cloudPlatform(CloudPlatform.GCP)
-        .region(GoogleRegion.DEFAULT_GOOGLE_REGION.name())
-        .schema(
-            new DatasetSpecificationModel()
-                .addTablesItem(
-                    new TableModel()
-                        .name(DATASET_TABLE_NAME)
-                        .addColumnsItem(
-                            new ColumnModel()
-                                .name(DATASET_COLUMN_NAME)
-                                .datatype(DATASET_COLUMN_TYPE))));
   }
 }
